[Priyanka] adding GreetingController to return JSON for different HTTP Methods
