+++ .idea/compiler.xml
@@ -6,13 +6,13 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="Greeting-App" />
+        <module name="GreetingApp" />
       </profile>
     </annotationProcessing>
   </component>
   <component name="JavacSettings">
     <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="Greeting-App" options="-parameters" />
+      <module name="GreetingApp" options="-parameters" />
     </option>
   </component>
 </project>
