+++ .idea/encodings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/Greeting-App/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/GreetingApp/src/main/java" charset="UTF-8" />
   </component>
 </project>
