+++ .idea/misc.xml
@@ -4,7 +4,7 @@
   <component name="MavenProjectsManager">
     <option name="originalFiles">
       <list>
-        <option value="$PROJECT_DIR$/Greeting-App/pom.xml" />
+        <option value="$PROJECT_DIR$/GreetingApp/pom.xml" />
       </list>
     </option>
   </component>
