--- Greeting-App/src/main/java/com/example/Greeting/App/Controller/GreetingController.java
@@ -1,40 +1,0 @@
-package com.example.Greeting.App.Controller;
-
-import org.springframework.web.bind.annotation.*;
-import java.util.HashMap;
-import java.util.Map;
-
-@RestController
-@RequestMapping("/greeting")
-public class GreetingController {
-
-    @GetMapping
-    public Map<String, String> getGreeting() {
-        Map<String, String> response = new HashMap<>();
-        response.put("message", "Hello from GET");
-        return response;
-    }
-
-    @PostMapping
-    public Map<String, String> postGreeting(@RequestBody Map<String, String> request) {
-        Map<String, String> response = new HashMap<>();
-        response.put("message", "Hello from POST");
-        response.put("received", request.get("name"));
-        return response;
-    }
-
-    @PutMapping
-    public Map<String, String> putGreeting(@RequestBody Map<String, String> request) {
-        Map<String, String> response = new HashMap<>();
-        response.put("message", "Hello from PUT");
-        response.put("updated", request.get("name"));
-        return response;
-    }
-
-    @DeleteMapping
-    public Map<String, String> deleteGreeting() {
-        Map<String, String> response = new HashMap<>();
-        response.put("message", "Hello from DELETE");
-        return response;
-    }
-}
