+++ GreetingApp/src/main/java/com/capgemini/GreetingApp/GreetingAppApplication.java
@@ -1,4 +1,4 @@
-package com.example.Greeting.App;
+package com.capgemini.GreetingApp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
